requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pandas
openpyxl
python-dotenv
requests